add trimester controller
